Fixed Update

+++ database/migrations/2024_02_06_012431_create_albums_table.php
@@ -7,13 +7,15 @@ use Illuminate\Support\Facades\Schema;
 return new class extends Migration {
     public function up(): void
     {
-        Schema::create('albums', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->string('nama_album');
-            $table->text('deskripsi');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('albums')) {
+            Schema::create('albums', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->string('nama_album');
+                $table->text('deskripsi');
+                $table->timestamps();
+            });
+        }
     }
 
     public function down(): void
