+++ database/migrations/2024_03_05_033457_create_komentarfoto_table.php
@@ -8,15 +8,17 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('komentarfoto', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('foto_id');
-            $table->foreign('foto_id')->references('id')->on('foto')->onDelete('cascade');
-            $table->unsignedBigInteger('user_id');
-            $table->foreign('user_id')->references('id')->on('users');
-            $table->text('isi_komentar');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('komentarfoto')) {
+            Schema::create('komentarfoto', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('foto_id');
+                $table->foreign('foto_id')->references('id')->on('foto')->onDelete('cascade');
+                $table->unsignedBigInteger('user_id');
+                $table->foreign('user_id')->references('id')->on('users');
+                $table->text('isi_komentar');
+                $table->timestamps();
+            });
+        }
     }
 
     public function down(): void
